requirements.txt
pyTelegramBotAPI>=4.12.0
GitPython>=3.1.30
docker>=6.1.2
python-dotenv>=1.0.0
aiohttp>=3.8.1
asyncio>=3.4.3
psutil>=5.9.0
requests>=2.28.0
markupsafe>=2.0.1 